upgraded Cargo.toml split program into 3 files: main.rs, encryption.rs and io.rs
