+++ Cargo.toml
@@ -6,6 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bcrypt = "0.15.0"
-clap = { version = "4.4.18", features = ["derive"] }
+bcrypt = "0.15.1"
+clap = { version = "4.5.3", features = ["derive"] }
 magic-crypt = "3.1.13"
